+++ desktop_version/src/Map.h
@@ -88,7 +88,7 @@ public:
     int roomdeaths[20 * 20];
     int roomdeathsfinal[20 * 20];
     static const int areamap[20 * 20];
-    int contents[40 * 30];
+    short contents[40 * 30];
     bool explored[20 * 20];
     int vmult[30];
